services/pyproject.toml
[project]
name = "services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.1.1",
    "fastapi>=0.119.0",
    "haystack-ai>=2.18.1",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.26.5",
    "pymupdf4llm>=0.0.27",
    "python-multipart>=0.0.20",
    "tiktoken>=0.12.0",
    "uvicorn>=0.37.0",
]
